mysql-sample/requirements.txt
google-auth
cloud-sql-python-connector[pymysql]
sqlalchemy
